requirements.txt
annotated-types==0.7.0
anyio==4.6.0
cassandra-driver==3.29.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.1
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.0
geomet==0.2.1.post1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jiter==0.5.0
openai==1.51.0
packaging==24.1
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==4.1.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
tomli==2.0.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6